+++ src/extensions/economy/generators/volcanicAshOperationsTypes.ts
@@ -1,8 +1,0 @@
-export interface VolcanicAshOperation {
-  i: number;
-  burgId: number;
-  marketId: number;
-  ashWorkers: number;
-  volcanicNeighborCount: number;
-  active: boolean;
-}